# Conflicts:
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.OperatorApp/ViewsAndViewModels/DrawerOperations/DrawerActivityInventoryViewModel.cs
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.OperatorApp/ViewsAndViewModels/DrawerOperations/DrawerActivityPickingViewModel.cs
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.OperatorApp/ViewsAndViewModels/SearchItem/ItemSearchViewModel.cs
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.OperatorApp/ViewsAndViewModels/WaitingLists/ListDetail/DetailListInWaitViewModel.cs
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.OperatorApp/ViewsAndViewModels/WaitingLists/ListsInWaitViewModel.cs
#	VerticalWarehouses/VerticalWarehouses.sln

+++ VerticalWarehouses/Machine Automation Service/MAS_DataLayer/DataLayerService.ShutterManualMovements.cs
@@ -7,13 +7,13 @@ namespace Ferretto.VW.MAS.DataLayer
     {
         #region Properties
 
-        public decimal Acceleration => this.GetDecimalConfigurationValue((long)ShutterManualMovements.Acceleration, (long)ConfigurationCategory.ShutterManualMovements);
+        public decimal Acceleration => this.GetDecimalConfigurationValue((long)ShutterManualMovements.Acceleration, ConfigurationCategory.ShutterManualMovements);
 
-        public decimal Deceleration => this.GetDecimalConfigurationValue((long)ShutterManualMovements.Deceleration, (long)ConfigurationCategory.ShutterManualMovements);
+        public decimal Deceleration => this.GetDecimalConfigurationValue((long)ShutterManualMovements.Deceleration, ConfigurationCategory.ShutterManualMovements);
 
-        public decimal FeedRateSM => this.GetDecimalConfigurationValue((long)ShutterManualMovements.FeedRate, (long)ConfigurationCategory.ShutterManualMovements);
+        public decimal FeedRateSM => this.GetDecimalConfigurationValue((long)ShutterManualMovements.FeedRate, ConfigurationCategory.ShutterManualMovements);
 
-        public decimal MaxSpeed => this.GetDecimalConfigurationValue((long)ShutterManualMovements.MaxSpeed, (long)ConfigurationCategory.ShutterManualMovements);
+        public decimal MaxSpeed => this.GetDecimalConfigurationValue((long)ShutterManualMovements.MaxSpeed, ConfigurationCategory.ShutterManualMovements);
 
         #endregion
     }
